+++ operator/tests/tutils/ais_cluster.go
@@ -25,8 +25,8 @@ const (
 	DefaultNodeImage     = "aistorage/aisnode:v3.28"
 	DefaultInitImage     = "aistorage/ais-init:v3.28"
 	DefaultLogsImage     = "aistorage/ais-logs:v1.0"
-	DefaultPrevNodeImage = "aistorage/aisnode:v3.27"
-	DefaultPrevInitImage = "aistorage/ais-init:v3.27"
+	DefaultPrevNodeImage = "aistorage/aisnode:v3.28-rc1"
+	DefaultPrevInitImage = "aistorage/ais-init:v3.28-rc1"
 	TestNSName           = "ais-op-test"
 	TestNSAnotherName    = "ais-op-test-other"
 )